+++ packages/react-router-dom/modules/BrowserRouter.js
@@ -1,7 +1,7 @@
 import warning from 'warning'
 import React from 'react'
 import PropTypes from 'prop-types'
-import createHistory from 'history/createBrowserHistory'
+import { createBrowserHistory as createHistory } from 'history'
 import Router from './Router'
 
 /**