+++ website/modules/animated/DecayAnimation.js
@@ -8,13 +8,13 @@
  *
  * @flow
  */
-'use strict';
+'use strict'
 
-var Animation = require('./Animation');
-var RequestAnimationFrame = require('./injectable/RequestAnimationFrame');
-var CancelAnimationFrame = require('./injectable/CancelAnimationFrame');
+var Animation = require('./Animation')
+var RequestAnimationFrame = require('./injectable/RequestAnimationFrame')
+var CancelAnimationFrame = require('./injectable/CancelAnimationFrame')
 
-import type { AnimationConfig, EndCallback } from './Animation';
+import type { AnimationConfig, EndCallback } from './Animation'
 
 type DecayAnimationConfigSingle = AnimationConfig & {
   velocity: number;
@@ -33,10 +33,10 @@ class DecayAnimation extends Animation {
   constructor(
     config: DecayAnimationConfigSingle,
   ) {
-    super();
-    this._deceleration = config.deceleration !== undefined ? config.deceleration : 0.998;
-    this._velocity = config.velocity;
-    this.__isInteraction = config.isInteraction !== undefined ? config.isInteraction : true;
+    super()
+    this._deceleration = config.deceleration !== undefined ? config.deceleration : 0.998
+    this._velocity = config.velocity
+    this.__isInteraction = config.isInteraction !== undefined ? config.isInteraction : true
   }
 
   start(
@@ -44,40 +44,40 @@ class DecayAnimation extends Animation {
     onUpdate: (value: number) => void,
     onEnd: ?EndCallback,
   ): void {
-    this.__active = true;
-    this._lastValue = fromValue;
-    this._fromValue = fromValue;
-    this._onUpdate = onUpdate;
-    this.__onEnd = onEnd;
-    this._startTime = Date.now();
-    this._animationFrame = RequestAnimationFrame.current(this.onUpdate.bind(this));
+    this.__active = true
+    this._lastValue = fromValue
+    this._fromValue = fromValue
+    this._onUpdate = onUpdate
+    this.__onEnd = onEnd
+    this._startTime = Date.now()
+    this._animationFrame = RequestAnimationFrame.current(this.onUpdate.bind(this))
   }
 
   onUpdate(): void {
-    var now = Date.now();
+    var now = Date.now()
 
     var value = this._fromValue +
       (this._velocity / (1 - this._deceleration)) *
-      (1 - Math.exp(-(1 - this._deceleration) * (now - this._startTime)));
+      (1 - Math.exp(-(1 - this._deceleration) * (now - this._startTime)))
 
-    this._onUpdate(value);
+    this._onUpdate(value)
 
     if (Math.abs(this._lastValue - value) < 0.1) {
-      this.__debouncedOnEnd({finished: true});
-      return;
+      this.__debouncedOnEnd({finished: true})
+      return
     }
 
-    this._lastValue = value;
+    this._lastValue = value
     if (this.__active) {
-      this._animationFrame = RequestAnimationFrame.current(this.onUpdate.bind(this));
+      this._animationFrame = RequestAnimationFrame.current(this.onUpdate.bind(this))
     }
   }
 
   stop(): void {
-    this.__active = false;
-    CancelAnimationFrame.current(this._animationFrame);
-    this.__debouncedOnEnd({finished: false});
+    this.__active = false
+    CancelAnimationFrame.current(this._animationFrame)
+    this.__debouncedOnEnd({finished: false})
   }
 }
 
-module.exports = DecayAnimation;
+module.exports = DecayAnimation