+++ website/modules/animated/injectable/RequestAnimationFrame.js
@@ -8,13 +8,13 @@
  *
  * @flow
  */
-'use strict';
+'use strict'
 
 var RequestAnimationFrame = {
   current: cb => global.requestAnimationFrame(cb),
   inject(injected) {
-    RequestAnimationFrame.current = injected;
-  },
-};
+    RequestAnimationFrame.current = injected
+  }
+}
 
-module.exports = RequestAnimationFrame;
+module.exports = RequestAnimationFrame